# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * POPNAME – Human Name Index API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface NameSummary {
  name: string;
  count: number;
  countries: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  meaning?: string | null;
  /** @nullable */
  gender?: string | null;
}

export type TrendingNameTrend = typeof TrendingNameTrend[keyof typeof TrendingNameTrend];


export const TrendingNameTrend = {
  rising: 'rising',
  falling: 'falling',
} as const;

export interface TrendingName {
  name: string;
  count: number;
  countries: number;
  changePercent: number;
  trend: TrendingNameTrend;
  sparkline?: number[];
}

export interface CountryCount {
  country: string;
  countryCode: string;
  count: number;
  percentage: number;
}

export interface NameDetail {
  name: string;
  count: number;
  countries: number;
  origin: string;
  meaning: string;
  gender: string;
  topCountries: CountryCount[];
  /** @nullable */
  changePercent?: number | null;
  sparkline?: number[];
}

export interface PopularityPoint {
  year: number;
  count: number;
  /** @nullable */
  rank?: number | null;
}

export interface DecadeNames {
  decade: number;
  names: string[];
}

export interface IndexStats {
  totalNamesIndexed: number;
  uniqueNames: number;
  countriesCovered: number;
  peopleAnalyzed: number;
  lastUpdated: string;
  dailyGrowth?: number;
  verifiedClaims?: number;
}

export interface ClaimInput {
  name: string;
  fullName: string;
  country: string;
  birthYear: number;
  /** @nullable */
  documentType?: string | null;
  /** @nullable */
  email?: string | null;
}

export type ClaimStatus = typeof ClaimStatus[keyof typeof ClaimStatus];


export const ClaimStatus = {
  pending: 'pending',
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface Claim {
  id: string;
  name: string;
  fullName?: string;
  country?: string;
  status: ClaimStatus;
  submittedAt: string;
}

export interface ForgeInput {
  baseNames?: string[];
  /** @nullable */
  meaning?: string | null;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  gender?: string | null;
  /**
     * e.g. classic, modern, mythological, nature
     * @nullable
     */
  style?: string | null;
}

export interface ForgedName {
  name: string;
  meaning: string;
  origin: string;
  isUnique: boolean;
  existingCount?: number;
  inspirations?: string[];
}

export interface ForgeResult {
  suggestions: ForgedName[];
}

export type SearchNamesParams = {
q: string;
limit?: number;
};

export type GetPopularNamesParams = {
limit?: number;
/**
 * @nullable
 */
country?: string | null;
};

export type GetTrendingNamesParams = {
period?: GetTrendingNamesPeriod;
limit?: number;
};

export type GetTrendingNamesPeriod = typeof GetTrendingNamesPeriod[keyof typeof GetTrendingNamesPeriod];


export const GetTrendingNamesPeriod = {
  '1m': '1m',
  '6m': '6m',
  '1y': '1y',
  '5y': '5y',
} as const;

export type GetDecliningNamesParams = {
period?: GetDecliningNamesPeriod;
limit?: number;
};

export type GetDecliningNamesPeriod = typeof GetDecliningNamesPeriod[keyof typeof GetDecliningNamesPeriod];


export const GetDecliningNamesPeriod = {
  '1m': '1m',
  '6m': '6m',
  '1y': '1y',
  '5y': '5y',
} as const;

export type GetRareNamesParams = {
limit?: number;
};