# lib/api-zod/src/generated/types/claimInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * POPNAME – Human Name Index API
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimInput {
  name: string;
  fullName: string;
  country: string;
  birthYear: number;
  /** @nullable */
  documentType?: string | null;
  /** @nullable */
  email?: string | null;
}