# lib/api-zod/src/generated/types/countryCount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * POPNAME – Human Name Index API
 * OpenAPI spec version: 0.1.0
 */

export interface CountryCount {
  country: string;
  countryCode: string;
  count: number;
  percentage: number;
}