# lib/api-zod/src/generated/types/forgeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * POPNAME – Human Name Index API
 * OpenAPI spec version: 0.1.0
 */

export interface ForgeInput {
  baseNames?: string[];
  /** @nullable */
  meaning?: string | null;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  gender?: string | null;
  /**
     * e.g. classic, modern, mythological, nature
     * @nullable
     */
  style?: string | null;
}