# lib/api-zod/src/generated/types/forgeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * POPNAME – Human Name Index API
 * OpenAPI spec version: 0.1.0
 */
import type { ForgedName } from './forgedName';

export interface ForgeResult {
  suggestions: ForgedName[];
}