# lib/api-zod/src/generated/types/nameSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * POPNAME – Human Name Index API
 * OpenAPI spec version: 0.1.0
 */

export interface NameSummary {
  name: string;
  count: number;
  countries: number;
  /** @nullable */
  origin?: string | null;
  /** @nullable */
  meaning?: string | null;
  /** @nullable */
  gender?: string | null;
}