# lib/db/src/schema/claims.ts
import { pgTable, text, serial, integer, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const claimsTable = pgTable("claims", {
  id: serial("id").primaryKey(),
  claimUuid: text("claim_uuid").notNull().unique(),
  name: text("name").notNull(),
  fullName: text("full_name").notNull(),
  country: text("country").notNull(),
  birthYear: integer("birth_year").notNull(),
  documentType: text("document_type"),
  email: text("email"),
  status: text("status").notNull().default("pending"),
  submittedAt: timestamp("submitted_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertClaimSchema = createInsertSchema(claimsTable).omit({ id: true, submittedAt: true });
export type InsertClaim = z.infer<typeof insertClaimSchema>;
export type Claim = typeof claimsTable.$inferSelect;